Upgrade code for namespaces Apply SilverStripe\Reports namespace
