+++ code/Report.php
@@ -1,9 +1,28 @@
 <?php
 
+namespace SilverStripe\Reports;
+
 use SilverStripe\ORM\ArrayList;
 use SilverStripe\ORM\SS_List;
 use SilverStripe\Security\Member;
 use SilverStripe\Security\Permission;
+use SilverStripe\Control\Controller;
+use SilverStripe\Core\ClassInfo;
+use SilverStripe\Forms\FieldList;
+use SilverStripe\Forms\LiteralField;
+use SilverStripe\Forms\FormAction;
+use SilverStripe\Forms\GridField\GridFieldConfig;
+use SilverStripe\Forms\GridField\GridFieldButtonRow;
+use SilverStripe\Forms\GridField\GridFieldPrintButton;
+use SilverStripe\Forms\GridField\GridFieldExportButton;
+use SilverStripe\Forms\GridField\GridFieldToolbarHeader;
+use SilverStripe\Forms\GridField\GridFieldSortableHeader;
+use SilverStripe\Forms\GridField\GridFieldDataColumns;
+use SilverStripe\Forms\GridField\GridFieldPaginator;
+use SilverStripe\Forms\GridField\GridField;
+use SilverStripe\Core\Convert;
+use SilverStripe\View\ViewableData;
+use ReflectionClass;
 
 /**
  * Base "abstract" class creating reports on your data.
@@ -32,8 +51,6 @@ use SilverStripe\Security\Permission;
  *
  * Right now, all subclasses of SS_Report will be shown in the ReportAdmin. In SS3 there is only
  * one place where reports can go, so this class is greatly simplifed from its version in SS2.
- *
- * @package reports
  */
 class SS_Report extends ViewableData
 {
@@ -71,9 +88,9 @@ class SS_Report extends ViewableData
      * @var array
      */
     public static $excluded_reports = array(
-        'SS_Report',
-        'SS_ReportWrapper',
-        'SideReportWrapper',
+        'SilverStripe\\Reports\\SS_Report',
+        'SilverStripe\\Reports\\SS_ReportWrapper',
+        'SilverStripe\\Reports\\SideReportWrapper',
     );
 
     /**
@@ -325,7 +342,7 @@ class SS_Report extends ViewableData
             new GridFieldPaginator()
         );
         $gridField = new GridField('Report', null, $items, $gridFieldConfig);
-        $columns = $gridField->getConfig()->getComponentByType('GridFieldDataColumns');
+        $columns = $gridField->getConfig()->getComponentByType('SilverStripe\\Forms\\GridField\\GridFieldDataColumns');
         $displayFields = array();
         $fieldCasting = array();
         $fieldFormatting = array();
@@ -424,123 +441,5 @@ class SS_Report extends ViewableData
     public function TreeTitle()
     {
         return $this->title();
-    }
-}
-
-/**
- * SS_ReportWrapper is a base class for creating report wappers.
- *
- * Wrappers encapsulate an existing report to alter their behaviour - they are implementations of
- * the standard GoF decorator pattern.
- *
- * This base class ensure that, by default, wrappers behave in the same way as the report that is
- * being wrapped.  You should override any methods that need to behave differently in your subclass
- * of SS_ReportWrapper.
- *
- * It also makes calls to 2 empty methods that you can override {@link beforeQuery()} and
- * {@link afterQuery()}
- *
- * @package reports
- */
-abstract class SS_ReportWrapper extends SS_Report
-{
-    protected $baseReport;
-
-    public function __construct($baseReport)
-    {
-        $this->baseReport = is_string($baseReport) ? new $baseReport : $baseReport;
-        $this->dataClass = $this->baseReport->dataClass();
-        parent::__construct();
-    }
-
-    public function ID()
-    {
-        return get_class($this->baseReport) . '_' . get_class($this);
-    }
-
-    ///////////////////////////////////////////////////////////////////////////////////////////
-    // Filtering
-
-    public function parameterFields()
-    {
-        return $this->baseReport->parameterFields();
-    }
-
-    ///////////////////////////////////////////////////////////////////////////////////////////
-    // Columns
-
-    public function columns()
-    {
-        return $this->baseReport->columns();
-    }
-
-    ///////////////////////////////////////////////////////////////////////////////////////////
-    // Querying
-
-    /**
-     * Override this method to perform some actions prior to querying.
-     */
-    public function beforeQuery($params)
-    {
-    }
-
-    /**
-     * Override this method to perform some actions after querying.
-     */
-    public function afterQuery()
-    {
-    }
-
-    public function sourceQuery($params)
-    {
-        if ($this->baseReport->hasMethod('sourceRecords')) {
-            // The default implementation will create a fake query from our sourceRecords() method
-            return parent::sourceQuery($params);
-        } elseif ($this->baseReport->hasMethod('sourceQuery')) {
-            $this->beforeQuery($params);
-            $query = $this->baseReport->sourceQuery($params);
-            $this->afterQuery();
-            return $query;
-        } else {
-            user_error("Please override sourceQuery()/sourceRecords() and columns() in your base report", E_USER_ERROR);
-        }
-    }
-
-    public function sourceRecords($params = array(), $sort = null, $limit = null)
-    {
-        $this->beforeQuery($params);
-        $records = $this->baseReport->sourceRecords($params, $sort, $limit);
-        $this->afterQuery();
-        return $records;
-    }
-
-
-    ///////////////////////////////////////////////////////////////////////////////////////////
-    // Pass-through
-
-    public function title()
-    {
-        return $this->baseReport->title();
-    }
-
-    public function group()
-    {
-        /** @skipUpgrade */
-        return $this->baseReport->hasMethod('group') ? $this->baseReport->group() : 'Group';
-    }
-
-    public function sort()
-    {
-        return $this->baseReport->hasMethod('sort') ? $this->baseReport->sort() : 0;
-    }
-
-    public function description()
-    {
-        return $this->baseReport->description();
-    }
-
-    public function canView($member = null)
-    {
-        return $this->baseReport->canView($member);
     }
 }
