+++ composer.json
@@ -1,29 +1,34 @@
 {
-	"name": "silverstripe/reports",
-	"type": "silverstripe-module",
-	"description": "Reports module for SilverStripe CMS",
-	"homepage": "http://silverstripe.org",
-	"license": "BSD-3-Clause",
-	"keywords": ["silverstripe", "cms", "reports"],
-	"authors": [
-		{
-			"name": "SilverStripe",
-			"homepage": "http://silverstripe.com"
-		},
-		{
-			"name": "The SilverStripe Community",
-			"homepage": "http://silverstripe.org"
-		}
-	],
-	"require": {
-		"silverstripe/framework": "^4.0"
-	},
-	"extra": {
-		"branch-alias": {
-			"dev-master": "4.0.x-dev"
-		}
-	},
-	"require-dev": {
-		"phpunit/PHPUnit": "~4.8"
-	}
+  "name": "silverstripe/reports",
+  "type": "silverstripe-module",
+  "description": "Reports module for SilverStripe CMS",
+  "homepage": "http://silverstripe.org",
+  "license": "BSD-3-Clause",
+  "keywords": ["silverstripe", "cms", "reports"],
+  "authors": [
+    {
+      "name": "SilverStripe",
+      "homepage": "http://silverstripe.com"
+    },
+    {
+      "name": "The SilverStripe Community",
+      "homepage": "http://silverstripe.org"
+    }
+  ],
+  "require": {
+    "silverstripe/framework": "^4.0"
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-master": "4.0.x-dev"
+    }
+  },
+  "autoload": {
+    "psr-4": {
+      "SilverStripe\\Reports\\": "code/"
+    }
+  },
+  "require-dev": {
+    "phpunit/PHPUnit": "~4.8"
+  }
 }
