+++ templates/SilverStripe/Reports/Includes/ReportAdmin_Content.ss
